+++ views/js/src/back/transaction/components/orderlines/OrderLinesTableActions.tsx
@@ -22,6 +22,7 @@ export default function OrderLinesTableActions({line, loading, shipLine, cancelL
     })), []);
 
     const isRefundable = (): boolean => line.refundableQuantity >= 1 && parseFloat(availableRefundAmount.value) > 0.0
+    const isCancelable = (): boolean => line.cancelableQuantity >= 1
 
     let shipButton = (
         <button
@@ -109,7 +110,7 @@ export default function OrderLinesTableActions({line, loading, shipLine, cancelL
                 cursor: (loading || line.cancelableQuantity < 1 || line.type === 'discount') ? 'not-allowed' : 'pointer',
                 opacity: (loading || line.cancelableQuantity < 1 || line.type === 'discount') ? 0.8 : 1,
             }}
-            onClick={() => isRefundable() && cancelLine([line])}
+            onClick={() => isCancelable() && cancelLine([line])}
             role="button"
         >
             <FontAwesomeIcon icon={!loading ? faTimes : faCircleNotch} spin={loading}/> {translations.cancel}