+++ changelog.md
@@ -4,6 +4,7 @@
 ## Changes in release 4.3.0 ##
 + Fixed custom url for carriers when no tracking number is provided
 + Fixed multishop configuration
++ Fixed order cancel button functionality in BO
 
 ## Changes in release 4.2.4 ##
 + Added optional custom order status for open bank transfer